pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sql-repair-env"
version = "2.0.0"
description = "RL environment where agents learn to diagnose and fix broken SQL queries"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.0",
    "duckdb>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "pydantic>=2.9.2",
    "openai>=1.51.0",
    "requests>=2.32.3",
    "pyyaml>=6.0.2",
]

[project.scripts]
server = "server.app:main"

[project.urls]
Homepage = "https://huggingface.co/spaces/iambhavyaa/sql-repair-env"
Repository = "https://github.com/bhavyasharma2/sql-repair-env"
